requirements.txt
pyvisa>=1.10.1
robotframework >= 2.8.7